postgres/requirements.txt
psycopg2
fastapi
uvicorn
python-dotenv